# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.ErrorResponse:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
  models.ExportRequest:
    properties:
      filters:
        $ref: '#/definitions/models.LeadSearchRequest'
      format:
        enum:
        - csv
        - excel
        type: string
      max_leads:
        maximum: 10000
        minimum: 1
        type: integer
    required:
    - format
    type: object
  models.ExportResponse:
    properties:
      created_at:
        type: string
      expires_at:
        type: string
      file_url:
        type: string
      format:
        type: string
      id:
        type: integer
      lead_count:
        type: integer
      status:
        type: string
    type: object
...
  models.LeadSearchRequest:
    properties:
      city:
        type: string
      country:
        type: string
      cuisineType:
        type: string
      hasEmail:
        type: boolean
      hasPhone:
        type: boolean
      hasSocialMedia:
        type: boolean
      hasWebsite:
        type: boolean
      industry:
        enum:
        - tattoo
        - beauty
        - barber
        - gym
        - restaurant
        - cafe
        - bar
        - bakery
        - dentist
        - pharmacy
        - massage
        - car_repair
        - car_wash
        - car_dealer
        - clothing
        - convenience
        - lawyer
        - accountant
        - spa
        - nail_salon
        type: string
      latitude:
        description: Radius search parameters
        maximum: 90
        minimum: -90
        type: number
      limit:
        maximum: 100
        minimum: 1
        type: integer
      longitude:
        maximum: 180
        minimum: -180
        type: number
      page:
        minimum: 1
        type: integer
      query:
        description: Full-text search
        type: string
      radius:
        minimum: 0
        type: number
      sortBy:
        description: Sorting
        enum:
        - newest
        - quality_score
        - distance
        - verified
        - relevance
        type: string
      specialties:
        items:
          type: string
        type: array
      sportType:
        type: string
      subNiche:
        type: string
      tattooStyle:
        type: string
      unit:
        enum:
        - km
        - miles
        type: string
      verified:
        type: boolean
    type: object
...
host: localhost:7890
info:
  contact:
    email: support@industrydb.io
    name: API Support
    url: https://industrydb.io/support
  description: Industry-specific business data API. Verified. Affordable.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://industrydb.io/terms
  title: IndustryDB API
  version: "1.0"
paths:
...
  /exports:
    get:
      description: Get paginated list of all exports created by the current user
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Items per page (default: 20, max: 100)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of exports with pagination
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List user exports
      tags:
      - Exports
    post:
      consumes:
      - application/json
      description: Create a new data export in CSV or Excel format with optional filters
      parameters:
      - description: Export configuration
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.ExportRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Export created successfully
          schema:
            $ref: '#/definitions/models.ExportResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "402":
          description: Usage limit exceeded
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create new export
      tags:
      - Exports
  /exports/{id}:
    get:
      description: Get detailed information about a specific export including status
        and download URL
      parameters:
      - description: Export ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Export details
          schema:
            $ref: '#/definitions/models.ExportResponse'
        "400":
          description: Invalid export ID
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Export not found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get export details
      tags:
      - Exports
  /exports/{id}/download:
    get:
      description: Download the generated CSV or Excel file for a specific export
      parameters:
      - description: Export ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/octet-stream
      responses:
        "200":
          description: Export file (CSV or Excel)
          schema:
            type: file
        "400":
          description: Invalid export ID
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Export not found or file unavailable
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Download export file
      tags:
      - Exports
...
securityDefinitions:
  ApiKeyAuth:
    description: API Key for programmatic access (Business tier only)
    in: header
    name: X-API-Key
    type: apiKey
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"